requirements.txt
dask[dataframe]
numpy==1.19.2
pandas
scikit-learn
tensorflow-gpu